# feat: Implement Core Web Vitals performance optimizations