--- a/package.json
+++ b/package.json
@@ -105,6 +105,7 @@
     "@radix-ui/react-tooltip": "^1.1.2",
     "@supabase/supabase-js": "^2.39.8",
     "@tailwindcss/typography": "^0.5.16",
+    "@tanstack/react-virtual": "^3.13.12",
     "@types/js-yaml": "^4.0.9",
     "@types/minimatch": "^5.1.2",
     "@vitejs/plugin-react": "^4.3.1",
@@ -142,6 +143,7 @@
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^1.0.0",
     "vite-imagetools": "^7.1.0",
+    "web-vitals": "^5.1.0",
     "zod": "^3.23.8",
     "zustand": "^5.0.3"
   },