--- a/src/components/features/repository/progressive-repo-view.tsx
+++ b/src/components/features/repository/progressive-repo-view.tsx
@@ -4,7 +4,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useTimeRangeStore } from "@/lib/time-range-store";
 import { useProgressiveRepoData } from "@/hooks/use-progressive-repo-data";
-import { useIntersectionLoader } from "@/hooks/use-intersection-loader";
+import { useLazyLoadData } from "@/hooks/use-intersection-observer";
 import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar";
 import { GitPullRequest, Users } from "lucide-react";
 
@@ -22,7 +22,7 @@ export function ProgressiveRepoView() {
   const timeRange = useTimeRangeStore((state) => state.timeRange);
   const [includeBots] = useState(false);
   
-  const { critical, full, enhancement } = useProgressiveRepoData(
+  const progressiveData = useProgressiveRepoData(
     owner,
     repo,
     timeRange,
@@ -33,8 +33,8 @@ export function ProgressiveRepoView() {
   const { 
     ref: chartRef, 
     data: chartData, 
-    isLoading: chartLoading 
-  } = useIntersectionLoader(
+    loading: chartLoading 
+  } = useLazyLoadData(
     async () => {
       // Simulate loading chart data
       await new Promise(resolve => setTimeout(resolve, 1000));
@@ -53,12 +53,12 @@ export function ProgressiveRepoView() {
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">
-              {critical.loading ? (
+              {!progressiveData.basicInfo ? (
                 <Skeleton className="h-8 w-20" />
               ) : (
                 <div className="flex items-center gap-2">
                   <GitPullRequest className="h-5 w-5 text-muted-foreground" />
-                  {critical.basicInfo?.prCount || 0}
+                  {progressiveData.basicInfo?.prCount || 0}
                 </div>
               )}
             </div>
@@ -71,12 +71,12 @@ export function ProgressiveRepoView() {
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">
-              {critical.loading ? (
+              {!progressiveData.basicInfo ? (
                 <Skeleton className="h-8 w-20" />
               ) : (
                 <div className="flex items-center gap-2">
                   <Users className="h-5 w-5 text-muted-foreground" />
-                  {critical.basicInfo?.contributorCount || 0}
+                  {progressiveData.basicInfo?.contributorCount || 0}
                 </div>
               )}
             </div>
@@ -88,7 +88,7 @@ export function ProgressiveRepoView() {
             <CardTitle className="text-sm font-medium">Top Contributors</CardTitle>
           </CardHeader>
           <CardContent>
-            {critical.loading ? (
+            {!progressiveData.basicInfo ? (
               <div className="flex -space-x-2">
                 {[...Array(5)].map((_, i) => (
                   <Skeleton key={i} className="h-8 w-8 rounded-full" />
@@ -96,7 +96,7 @@ export function ProgressiveRepoView() {
               </div>
             ) : (
               <div className="flex -space-x-2">
-                {critical.basicInfo?.topContributors.map((contributor, i) => (
+                {progressiveData.basicInfo?.topContributors.map((contributor: any, i: number) => (
                   <Avatar key={contributor.id} className="h-8 w-8 border-2 border-background">
                     <AvatarImage 
                       src={`${contributor.avatar_url}?s=64`} 
@@ -113,13 +113,13 @@ export function ProgressiveRepoView() {
       </div>
 
       {/* Stage 2: Full Data - Render when available */}
-      {(full.loading || full.stats) && (
+      {(progressiveData.stats.loading || progressiveData.stats.pullRequests.length > 0) && (
         <Card>
           <CardHeader>
             <CardTitle>Repository Activity</CardTitle>
           </CardHeader>
           <CardContent>
-            {full.loading ? (
+            {progressiveData.stats.loading ? (
               <div className="space-y-2">
                 <Skeleton className="h-4 w-full" />
                 <Skeleton className="h-4 w-3/4" />
@@ -128,11 +128,11 @@ export function ProgressiveRepoView() {
             ) : (
               <div>
                 <p className="text-sm text-muted-foreground">
-                  {full.stats?.pullRequests.length || 0} pull requests in the last {timeRange} days
+                  {progressiveData.stats?.pullRequests.length || 0} pull requests in the last {timeRange} days
                 </p>
-                {full.lotteryFactor && (
+                {progressiveData.lotteryFactor && (
                   <p className="text-sm mt-2">
-                    Lottery Factor: {full.lotteryFactor.riskLevel} ({full.lotteryFactor.topContributorsPercentage.toFixed(1)}% from top {full.lotteryFactor.topContributorsCount} contributors)
+                    Lottery Factor: {progressiveData.lotteryFactor.riskLevel} ({progressiveData.lotteryFactor.topContributorsPercentage.toFixed(1)}% from top {progressiveData.lotteryFactor.topContributorsCount} contributors)
                   </p>
                 )}
               </div>
@@ -142,13 +142,13 @@ export function ProgressiveRepoView() {
       )}
 
       {/* Stage 3: Enhancement Data - Background loading */}
-      {(enhancement.loading || enhancement.directCommits) && (
+      {progressiveData.directCommitsData && (
         <Card>
           <CardHeader>
             <CardTitle>Direct Commits</CardTitle>
           </CardHeader>
           <CardContent>
-            {enhancement.loading ? (
+            {!progressiveData.stageProgress.enhancement ? (
               <div className="space-y-2">
                 <Skeleton className="h-4 w-full" />
                 <Skeleton className="h-4 w-3/4" />
@@ -156,11 +156,11 @@ export function ProgressiveRepoView() {
             ) : (
               <div>
                 <p className="text-sm text-muted-foreground">
-                  {enhancement.directCommits?.hasYoloCoders ? `${enhancement.directCommits.yoloCoderStats.length} YOLO coders detected` : 'No YOLO coders'}
+                  {progressiveData.directCommitsData?.hasYoloCoders ? `${progressiveData.directCommitsData.yoloCoderStats.length} YOLO coders detected` : 'No YOLO coders'}
                 </p>
-                {enhancement.directCommits?.yoloCoderStats && enhancement.directCommits.yoloCoderStats.length > 0 && (
+                {progressiveData.directCommitsData?.yoloCoderStats && progressiveData.directCommitsData.yoloCoderStats.length > 0 && (
                   <p className="text-sm mt-1">
-                    Top YOLO coder: {enhancement.directCommits.yoloCoderStats[0].login} ({enhancement.directCommits.yoloCoderStats[0].directCommitPercentage.toFixed(1)}% direct commits)
+                    Top YOLO coder: {progressiveData.directCommitsData.yoloCoderStats[0].login} ({progressiveData.directCommitsData.yoloCoderStats[0].directCommitPercentage.toFixed(1)}% direct commits)
                   </p>
                 )}
               </div>