… src/hooks/use-progressive-repo-data.ts
… src/lib/web-vitals-monitoring.ts