@bdougie 2025-08-07
## Performance Targets Achieved

- ✅ **LCP < 2.5 seconds** - Progressive loading ensures critical content appears quickly
- ✅ **INP < 200ms** - Virtualization and lazy loading reduce interaction delays  
- ✅ **CLS < 0.1** - Skeleton screens and dimension preservation prevent layout shifts